# openapi.yaml
openapi: 3.0.0
info:
  title: AI Screening API для Хакатона
  description: |
    Сервис первичного скрининга кандидатов.
    — Базовая форма → создание кандидата и сессии диалога.
    — Пошаговый чат с уточняющими вопросами.
    — Финальный скоринг + рекомендация.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Локальный Go-бэкенд (см. backend/)
paths:
  /candidates:
    post:
      summary: Создать кандидата (базовая форма + старт сессии)
      operationId: createCandidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateCreateRequest'
      responses:
        '201':
          description: Сессия диалога создана, возвращаются первые вопросы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogStartResponse'
        '400':
          description: Ошибка валидации
...
components:
  schemas:
...
    CandidateCreateRequest:
      type: object
      required:
        - position
        - grade
        - about
        - education
        - resume_url
      properties:
        candidate_id:
          type: string
          description: Опционально, если клиент хочет задать свой UUID. Иначе сервер сгенерирует.
        position:
          type: string
          example: "GO разработчик"
        grade:
          type: string
          enum: [Intern, Junior, Middle, Senior, Lead]
          example: "Middle"
        about:
          type: string
          description: "Немного о себе, ключевые достижения"
          example: "3 года писал на Go, занимался высокими нагрузками"
        education:
          type: string
          example: "МГТУ им. Баумана, Прикладная информатика, 2022"
        resume_url:
          type: string
          format: uri
          description: Ссылка на PDF или текстовый файл резюме (может быть пустой, если загружаете текст отдельно)
          example: "https://example.com/resume.pdf"
...
    DialogStartResponse:
      type: object
      required:
        - session_id
        - status
        - questions
        - max_questions
      properties:
        session_id:
          type: string
          description: ID сессии диалога
          example: "sess_12345"
        status:
          type: string
          enum: [in_progress]
        questions:
          type: array
          description: Первые вопросы (обычно 2 шт., чтобы сразу начать диалог)
          items:
            $ref: '#/components/schemas/QuestionItem'
        max_questions:
          type: integer
          description: Общее число запланированных вопросов (чтобы пользователь знал прогресс)
          example: 5
...
    QuestionItem:
      type: object
      required:
        - question_id
        - text
      properties:
        question_id:
          type: integer
        text:
          type: string
...